# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Janus 2.0 Security Clearance API",
    "version": "2.0.0",
    "description": "API for managing personnel, organizations, and three-tier access control in high-security environments",
    "contact": {
      "name": "Janus Development Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:15520/api",
      "description": "Development server"
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "Login with username and password",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 50
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "token": {"type": "string"},
                        "user_id": {"type": "string"},
                        "role": {"type": "string"}
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
...
  },
...
}